Rename instructionsComponentsTest to MarkdownInstructionsTest and move file
